refactor : #137 resaleScheduler Cancel 부분 리팩토링
